refactor(site-docs): redesign pricing page with extracted data layer
Extract tier data into src/data/pricing-tiers.ts and hero into
PricingHero component. Modernize TierCard visuals with refined
typography, glass-like borders, and hover interactions. Pricing
page is now a 22-line thin composition layer.
